Struct mongo_driver::client::ClientPool

pub struct ClientPool {
    // some fields omitted
}

Pool that allows usage of clients out of a single pool from multiple threads.

Use the pool to pop a client and do operations. The client will be automatically added back to the pool when it goes out of scope.

This client pool cannot be cloned, but it can be use from different threads by using an Arc. Clients cannot be shared between threads, pop a client from the pool for very single thread where you need a connection.

Methods

impl ClientPool

fn new(uri: Uri, ssl_options: Option<SslOptions>) -> ClientPool

Create a new ClientPool with that can provide clients pointing to the specified uri. The pool will connect via SSL if you add ?ssl=true to the uri. You can optionally pass in SSL options to configure SSL certificate usage and so on.

fn get_uri(&self) -> &Uri

Get a reference to this pool's Uri.

fn pop(&self) -> Client

Retrieve a client from the client pool, possibly blocking until one is available.